requirements.txt
Django==1.7.6
Pillow==2.7.0
PyJWT==0.4.3
PyYAML==3.11
argparse==1.2.1
awsebcli==3.1.2
boto==2.36.0
braintree==3.11.1
cement==2.4.0
django-allauth==0.19.1
django-appconf==1.0.1
django-countries==3.2
django-geoposition==0.2.2
django-payments==0.6.4
django-storages==1.1.8
jmespath==0.6.1
oauthlib==0.7.2
psycopg2==2.6
pycrypto==2.6.1
python-dateutil==2.4.1
python-openid==2.2.5
requests==2.5.3
requests-oauthlib==0.4.2
six==1.9.0
stripe==1.21.0
suds-jurko==0.6
wsgiref==0.1.2
